PHYSICS_SCALE, TASK 6/7 — Finish measured scale regression checks
Why:
The normal performance entry point now exercises all ten required workloads.
Final repeated measurements meet every matched acceptance threshold. GCC's
remaining enum/integer diagnostic still prevents its portable sanitizer lane.

Ownership:
Physics retains authority over collision-path flags; the explicit conversion
preserves the existing integral operation. Performance tooling owns the scale
acceptance evidence. No state, arithmetic, baseline or allocator policy moves.

What:
Make the existing speculative-disabled enum conversion to unsigned explicit
before the conditional operation. Ask Ninja to collect independent compilation
failures in one run with -k 0; its nonzero failure and pipefail behavior remain
enforced. Accept PS5 at 6/7 and portfolio 144/151 with the exact gate outcomes.

Validation:
PASS: Debug/Profile ready builds, compiler source-design check (one file/three
contexts), and unchanged staged Physics 0/repeat/1/4 matrix. The rebuilt Windows
Profile .text is byte-identical to the measured producer: 3,749,376 bytes,
SHA-256 c4120e4b3943613503ecf1ba29bd6072d772907c67cecfe326b1cb54c9b69231.
All 30 final measurements pass: 2000 2.8480->1.3858ms (51.3% faster), sleepy
5000 12.4594->0.9420ms (92.4%), gravity1024 3.9263->3.2905ms (16.2%), and
joints320 2.1987->1.3158ms (40.2%). All other workloads improve 3.7%-33.8%.
The normal perf gate passes ten new workload checks, negative controls, both
absolute budgets, allocation guards and causal-inspection cost proof. Its
overall exit remains 7 on sixteen legacy UI/frame/memory relative failures;
the preserved pre-change Profile reproduces every same failure category in
both workloads. No legacy performance limit is weakened. All 15 native
scene-transition prediction cases and the remaining Skarness state/query/
render/playback/shortcut constituents pass. Clang warning-clean, ASan and
UBSan pass hosted run 34690480793; GCC/TSan will rerun after this repair.

Baselines/Artifacts:
Every baseline remains unchanged (136 tracked baseline/golden paths checked).
Artifacts: TestOutput/physics-scale-optimization/final-portable/,
final-code-equivalence.json, legacy-perf-before/, remaining-local-validation.json,
staged-enum-physics.log and compiler logs. The final metrics bind a preserved
producer; the compile-only conversion produces exactly the same machine code.
PS6 remains open for final portable results and cumulative handoff.

Review:
Independent review is clean for the complete optimization, repaired memory
expectations and bounded integer sorting repair. Another review is not required
for making the existing integral conversion explicit or retaining all compiler
diagnostics in a failing build. Final Windows unit rerun is also in progress.

+++ SkullbonezSource/Physics/Stages/PhysicsMotionEligibilityStage.cpp
@@ -227,7 +227,7 @@ void PhysicsMotionEligibilityStage::Run( const PhysicsBodyStore& bodyStore,
         const int b = joint.BodyBIndex( bodyStore );
         if ( a >= 0 && b >= 0 && a != b && a < modelCount && b < modelCount )
         {
-            const uint8_t path = PhysicsMotionEligibilityArticulated | ( speculativeEnabled ? 0u : PhysicsMotionEligibilitySpeculativeDisabled );
+            const uint8_t path = PhysicsMotionEligibilityArticulated | ( speculativeEnabled ? 0u : static_cast<unsigned>( PhysicsMotionEligibilitySpeculativeDisabled ) );
             m_collisionPathState[a] = path;
             m_collisionPathState[b] = path;
         }
